requirements.txt
dash-mantine-components
dash-pdf
dash-iconify
pandas
gunicorn
